+++ src/components/header/header.jsx
@@ -2,12 +2,10 @@ import React from 'react';
 
 import styles from './header.less';
 
-export default class Status extends React.Component {
-  render() {
-    return (
-      <div className={styles['container']}>
-        <div className={styles['title']}>DJ Lama</div>
-      </div>
-    );
-  }
+export default function Status() {
+  return (
+    <div className={styles['container']}>
+      <div className={styles['title']}>DJ Lama</div>
+    </div>
+  );
 }