+++ src/components/status/status.jsx
@@ -2,13 +2,11 @@ import React from 'react';
 
 import styles from './status.less';
 
-export default class Status extends React.Component {
-  render() {
-    return (
-      <div className={styles['container']}>
-        <div className={styles['page-title']}>Status page</div>
-        <img className={styles['logo']} src='public/resources/logo.svg'/>
-      </div>
-    );
-  }
+export default function Status() {
+  return (
+    <div className={styles['container']}>
+      <div className={styles['page-title']}>Status page</div>
+      <img className={styles['logo']} src="public/resources/logo.svg" alt="DJ Lama logo" />
+    </div>
+  );
 }