+++ webpack.client.prod.config.js
@@ -1,13 +1,6 @@
 const webpack = require('webpack');
-const path = require('path');
 
-module.exports = {
-  entry: './src/client.js',
-  output: {
-    path: path.join(__dirname, 'build/public'),
-    publicPath: '/',
-    filename: 'client.js'
-  },
+const config = {
   plugins: [
     new webpack.optimize.DedupePlugin(),
     new webpack.optimize.UglifyJsPlugin({
@@ -19,24 +12,13 @@ module.exports = {
     }),
     new webpack.DefinePlugin({
       'process.env': {
-        'NODE_ENV': JSON.stringify('production')
+        NODE_ENV: JSON.stringify('production')
       }
     })
-  ],
-  module: {
-    loaders: [{
-      test: /\.jsx?$/,
-      exclude: /node_modules/,
-      loader: 'babel-loader',
-      query: {
-        presets: ['es2015', 'react']
-      }
-    }, {
-      test: /\.less$/,
-      loader: 'style-loader!css?sourceMap&modules&localIdentName=[hash:base64:15]!less'
-    }]
-  },
-  resolve: {
-    extensions: ['', '.js', '.jsx']
-  }
+  ]
+};
+
+module.exports = {
+  config,
+  cssHash: '[hash:base64:15]'
 };